+++ 02_snatch_metadata/params_snatch.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+@date: 2024
+@author: Samantha Pendleton
+@GitHub: github.com/sap218/jabberwocky
+"""
+
+####################################################
+#
+# PARAMETERS FOR BANDERSNATCH
+#
+# Any completed fields below are an example for running the test data
+#
+####################################################
+
+is_this_a_test = True
+
+####################################################
+
+dir_output = "output/"
+
+files_location = "input/"
+
+ontology_name = ""
+
+ontology_tags_name = "ontology_tags"
+
+classes_of_interest = ""
+
+####################################################
+
+# End of script