+++ 03_catch_text/params_catch.py
@@ -1,0 +1,49 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+@date: 2024
+@author: Samantha Pendleton
+@GitHub: github.com/sap218/jabberwocky
+"""
+
+####################################################
+#
+# PARAMETERS FOR CATCH
+#
+# Any completed fields below are recommendations or examples
+#
+####################################################
+
+is_this_a_test = True
+
+####################################################
+
+dir_output = "output/"
+
+files_location = "input/"
+
+stopWord_filter_level = "" # "light" or "heavy" or "none"
+
+file_corpus = ""
+
+plotWORDCLOUD = True
+if plotWORDCLOUD:
+    plotWORDCLOUDcolormap = ""
+
+stop_here = True # If True, user does not need to proceed with the rest of this
+stop_here = False # If False, user should provide the rest of the parameters
+
+#########################
+
+file_words_of_interest = ""
+
+output_format = "" # ["wtags","grep","invertedgrep"]
+output_style = "" # original formatted
+
+plotCYANNOTATOR = True
+if plotCYANNOTATOR:
+    plotCYANNOTATORhighlightcolour = "" 
+
+####################################################
+
+# End of script